src/api/serializers.py
from rest_framework import serializers
from web.models import User


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'avatar_url')