requirements.txt
Django==1.7.1
django-mathfilters==0.3.0
djangorestframework==3.0.2
drf-nested-routers==0.9.0
uWSGI==2.0.8
psycopg2==2.5.4
dj-database-url==0.3.0
dj-static==0.0.6
static==1.1.1
gunicorn==19.1.1
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.2.0